refactor(dosing): extract a neutral crate::dosing module out of ode/predictions
Dose resolution (`resolve_subject_doses{,_with}`, the #324 single source of truth for
modeled-`RATE` doses) and the SS-equilibration primitives (`SS_EQUILIBRATION_CYCLES`,
`SS_EQUILIBRATION_TOL`, `SS_TAIL_REL_FLOOR`, `ss_cycle_converged`, `SsStopTracker`, and the
test-only cycle recorder) lived inside `ode/predictions.rs`, forcing `pk/`, `sens/`, and
`api` to depend *upward* on `ode/` merely to resolve doses or share the SS convergence
policy. Moved them (verbatim) into a new neutral leaf module `crate::dosing`.

- `ode::predictions` re-exports every moved symbol (`pub(crate) use crate::dosing::{..}`),
  so its 7 internal resolve callers + `equilibrate_ss_state` compile unchanged and every
  historical `crate::ode::{resolve_subject_doses, predictions::*}` path still resolves.
- Repointed the ~30 external call sites in `pk/{mod,event_driven}`, `sens/{propagate,
  ode_provider}`, and `api` from `crate::ode::…` to `crate::dosing::…`.
- Collapsed two duplicate `const … = 50` shadows (`event_driven::EVENT_DRIVEN_SS_EQUILIBRATION_CYCLES`,
  `propagate::SS_EQUILIBRATION_CYCLES`) onto the canonical `crate::dosing::SS_EQUILIBRATION_CYCLES`
  (50 == 50, bit-identical). The predictions test-mod's direct thread-local reads go through
  the `last_ss_equilibration_cycles()` accessor.

Pure code motion — no f64 arithmetic reordered. Verified: cargo test --lib => 2516 passed,
0 failed (ss_cycle_converged / ss_equilibration_early_stop / resolve-dose paths green).
